remove 0x
