+++ src/constants/contracts.ts
@@ -1,8 +1,3 @@
-// 0x
-export const zeroExRouterAddress = '0xDef1C0ded9bec7F1a1670819833240f027b25EfF'
-export const zeroExRouterOptimismAddress =
-  '0xdef1abe32c034e558cdd535791643c58a13acc10'
-
 // Index
 export const DebtIssuanceModuleAddress =
   '0x69a592D2129415a4A1d1b1E309C17051B7F28d57'
