+++ src/lib/hooks/use-best-quote/utils/flashmint.ts
@@ -5,8 +5,7 @@ import { providers, utils } from 'ethers'
 import { Token } from '@/constants/tokens'
 import { IndexRpcProvider } from '@/lib/hooks/use-wallet'
 import { displayFromWei } from '@/lib/utils'
-import { getConfiguredZeroExApi } from '@/lib/utils/api/zeroex'
-import { getNetworkKey } from '@/lib/utils/api/zeroex-utils'
+import { getConfiguredZeroExApi, getNetworkKey } from '@/lib/utils/api/zeroex'
 import { getFullCostsInUsd, getGasCostsInUsd } from '@/lib/utils/costs'
 import { getFlashMintGasDefault } from '@/lib/utils/gas-defaults'
 import { GasEstimatooor } from '@/lib/utils/gas-estimatooor'
